If input file is JSON, code becomes much simpler
We no longer have any need for all the complex parsing code. I've also removed the merging code as that's part of the parsing, but merging defaults will still be required - but it's not gone as we have git.

We also do not need the config_loader test, as it only tested whether the yaml was parsed correctly.