Remove translate_yaml_input method
This is a precursor to cleaning up the code. A lot of work is being done many times which is hidden because it was pushed out into this method which in turn was called a few times. We are making the code more procedural but this is an interim step on the way to clearer, cleaner code :)

Also changing the tests that called this method as in fact they were also duplicating work.