Strip overly verbose input/output VSE data into files
The purpose of the LoadBalancerService generator is to take
viable input configurations and convert them to the exact data
structure returned by the get_edge_gateway Fog call. We need to be
confident that this process is working in entirety.

However, unlike the Firewall and Nat checks, the output format is
extremely verbose, so would massively clutter up the test file.

This commit pulls the current single HTTP load balancer test input/output
data out into separate serialized YAML files.