Jenkins to run one set of integration tests
As per other vCloud Tools, we want jenkins to run the unit tests, features and integration tests, and if all that passes, to publish (this commit doesn't do the publication, that is a subsequent commit). Here, we run the integration tests against one environment (the job in Jenkins will determine which one) as that should be sufficient. However, we have access to other environments, so we also require a jenkins_integration_tests script for just running the integration tests.