Fix commented out unit test
This failed when ConfigLoader was moved to Core and was commented out in the interests of finishing that story b694077cb6. This comments it back in and fixes the issues.

It would be much easier to write tests if the config was injected rather than a new config loader, but that would require a refactor of the bin file (because the config would have to be loaded there and that file is untested, and would need to be refactored to be testable) so it's a story for another day.