Add ability to authenticate with tokens
This functionality was added to vCloud Core after I began pulling out the vCloud Launcher, so the version bump and the instructions were added to vCloud Tools and are not present in vCloud Launcher. I was going to cherry-pick the commits from the vCloud Tools PR https://github.com/alphagov/vcloud-tools/pull/178 into this repo, but given the conflicts with Jeweller etc and the fact that the changes are very minor, it is simpler just to reapply them here.

Also update CHANGELOG but no need to bump version yet.