Bump vCloud Core version to use progress bar fix
Our supplier's update to 5.5 exposed an issue where the progress bar fell over when there was no progress to report. This was fixed in https://github.com/fog/fog/pull/2837 and released in Fog 1.22.0, so this commit bumps the version of vCloud Core required. It also bumps the vCloud Launcher version because the earlier version will not work against a 5.5 environment even if you specify using the 5.1 API.