Reinstate deleted tests
These tests were removed and one of them reimplemented as a unit test in 8cd8df57f. However, the second one was not reimplemented. This test demonstrates that when customization fails, the VM orchestrator leaves the VM in a half-finished state with no error message, i.e. it fails silently to do the rest of the customization. This test was highlighting that bug.

It was not reimplemented as a unit test because the plan was to fix the bug as the next piece of work - however, that person was then off sick and then moved on to another team. Reinstating it here so that at least it is tested.

The story to fix the issue https://www.pivotaltracker.com/n/projects/790363/stories/63207274.