Create `integration` task as alias for `:all`
All the other vCloud Tools have an `integration` rake task. Because the full integration tasks for launcher take a very long time, we have a `integration:quick` and an `integration:all` task.

This commit creates an `integration` task that is an alias for `integration:all`, so that a user new to this tool but familiar with the others can run tests. If they find the tests take too long they can investigate further (perhaps by reading the README) to find the `integration:quick` task.