Use pessimistic version dependency for vcloud-core
Use Ruby's pessimistic operator[1] for the version dependency on the
vcloud-core gem, which prevents us from installing a version of
vcloud-core with a greater minor version than the one specified.

In this case, only versions less than 0.1.0 will be installed.

This reinforces our policy of semantic versioning, whereby a minor
version bump may signify that changes in that versions are not backwards
compatible.

[1]: http://guides.rubygems.org/patterns/#declaring_dependencies