Suppress Fog progress bar in quiet mode
The only ways of suppressing the progress bar in the
current release of Fog are:

* Not be writing to a terminal
* Setting the :vcloud_director_show_progress option on each Fog connection created.
* Setting the 'vcloud_director_show_progress' FOG_RC credential option to false.

Not writing to a terminal (or faking this) is not sensible -- we're a CLI tool first and foremost.

Setting the option on each connection is problematic in the current code-base. We do
not use a single Fog connection object throughout, so would need to either:

* ensure that the --quiet option is passed throughout our code
* update our code to DI a single Fog connection object from the top level

The FOG_RC credentials option can be overridden by writing to Fog.credentials, as documented
in https://github.com/fog/fog-core/blob/master/lib/fog/core/credentials.rb#L6-L24

This, whilst a little non-intuitive, is useful to us as it sets the option globally:
if we create multiple Fog connection objects, they all will adopt this behaviour.