Stub ::Fog.credentials call disabling progress bar
The previous commit broke the unit tests, since we were talking to Fog
proper to set the credentials option to disable the progress bar.

Our setting of FOG_CREDENTIAL to a bogus value, which is present to
protect us from accidentally connecting to a real organisation,  picked
this up.

As discussed in the previous commit message, setting the Fog.credentials
configuration option is the only useful method we have at our disposal
to toggle display of the progress bar.

However, since this is a unit test we must stub out this call to the
external Fog library, and just ensure that we are setting the option we
know is required.