Change network parameter names for consistency
e.g. from `network1_ip` to `network_1_ip`, for consistency with other parameter names.

This commit also bumps the version of vCloud Tools Tester required - it's not necessary at this stage as this is not yet using vCloud Tools Tester, but the reason for requiring that version will be that is the version that includes the updated network names so committing this here for clarity.