Add template for the testing parameters & doc
This commit also adds the actual parameters file to `.gitignore` (to join quite a lot of cruft! A story for another day). The idea is that you copy the template file and put your actual parameters in there, so no danger of committing your environment details accidentally. As explained in the documentation.

This commit doesn't remove the environment variables documentation from the README as at this point, that is still how the integration tests are run.