Use vCloud Tools Tester to get parameters
This changes the way you run the tests (see the README change) but means that all tests can be run using a config file that we can share. It also means the only environment variable we need to set it is the FOG_CREDENTIAL so that removes the issue of parameters from one test interacting with those from another, and it also makes explicit the test parameters we are using.