Remove outer module block from unit tests
So that our tests aren't extending and implicitly namespaced into the
`Vcloud` module. This will make it more obvious where we're crossing the
boundaries of different projects. It will also instantiate the class given
(without any args) and make it available as `subject` which may allow us to
DRY up some of our tests.