Update Jenkins files to get parameters from GHE
At the moment the Jenkins jobs export a ton of environment variables which are then used in the tests. After the change to use vCloud Tools Tester is merged, the tests will be expecting input from a config yaml file.

This change clones the repository in which I have saved the config file we can use, copies the file to the correct place and then removes the repository. The removal is because although most of our Jenkins jobs have 'Delete workspace before build starts' checked, I would prefer to rely as little as possible on Jenkins config.

Having said that, this change requires that the Jenkins jobs, as well as exporting the variables required for the fog credentials, now need to export FOG_CREDENTIAL=[the name of the config block in the config file]. I have added that to both the vCloud Launcher Jenkins jobs. Should we add it to `tools/fog_credentials` instead?

This will need updating when we address https://www.pivotaltracker.com/story/show/71594524