Clean workspace before starting
Deleting the Gemfile.lock will mean jenkins has to resolve all dependencies again which is good from a testing perspective. In any case, we have the option Jenkins option 'Delete workspace before build starts' checked. This is just to make this explicit in code so that we don't depend on config in Jenkins.

Git clean removes untracked files - for the same reason above. Putting this also means we no longer need to explicitly remove the checked out config directory.