Use subject in VappOrchestrator test
To match style of the other unit tests. We need to redefine subject we want
the class rather than an instance of the class (which is what rspec gives us
from the top-level describe) in order to test the class methods. I wonder if
this should be a module instead?