Rename methods that return config/vApp/VM schemas
These methods are already namespaced to the class that's defining them. The
prefixes of `provision_` and `customize_` seem rather arbitrary and don't
really describe the contents.

These are about to be refactored to different locations where I'll just call
them `VAPP_SCHEMA` and `VM_SCHEMA`. Calling out this rename in preperation
for that.