Remove unnecessary tests
This commit reinstates the removal of some of the metadata, as per 6a957dec. This speeds up the tests by about a minute.

It also removes the tests that prove a bug whereby the customisation fails silently. It turns out that this has been fixed in the interim (so the test wasn't testing what was meant to be tested). The continuing customization after selecting the default storage profile is tested in a unit test.