Remove test for nonsense storage profile
This test was for the condition where we pass a storage profile that doesn't exist and try and provision a VM with that storage profile. Previously, this would use the default, fail silently, ignore the rest of the customisation and then continue with the next VM.

Now, this errors. I was thinking of adding a test for this specific funcionality, i.e. pass a nonsense storage profile, VM errors out on provisioning, but this is actually just an example of the case where the configuration is wrong which is tested elsewhere, so I'm not sure it's worth adding a specific test here.Removing this test also shaves another couple of minutes off this integration test.