Refactor: use Trollop for arg processing. Remove Cli class.
The Cli class is a lot of code and specs around optparser. Trollop is
far more compact and we can achieve what Cli was doing in significantly
less code.

Trollop symbolises keys in the opts hash it generates, unlike our
Cli/optparser class, so some small changes to follow that.