Bump dependency on vCloud Core version without fog
I did this to avoid later trouble that might be caused by updating vCloud Core to >= 0.10.0 without updating vCloud Tools Tester to >= 0.2.0, as there is a circular dependency. SO I thought I was doing a favour to later developers. However, it turns out that there was a stray inclusion of `vcloud/fog` here, so it's a good job I did this, and actually I'm just finishing the job I started :)