Run a post-launch-cmd on VAPP launch
In order to trigger actions when a VM is launched, we should be able to
specify an arbitrary script to run which receives information about the
VAPP via environment variables and is capable of executing further
actions (such as notifying other systems that the VM has been created).

This commit adds the -p flag which will run the executable specified on VAPP
launch. It will pass the entire vapp_definition hash as the
VAPP_DEFINITION environment variable and the script will need to process
that for further information.

This is for [#66416966]