Add IndependentDiskOrchestrator class & tests
In the interests of not cluttering up the VmOrchestrator class,
I've added an IndependentDiskOrchestrator class to handle
translation of the :independent_disks config subsection into input
to the Vcloud::Core::IndependentDisk class.

This will then be tied into the VmOrchestrator class, in the same
manner that VappOrchestrator calls VmOrchestrator.

This separation also allows for independent management of Independent Disks.