Add support for :independent_disks config section
:independent_disks is a list of hashes giving the :name of each independent disk
to attach. The vDC for the disk must come from the VM it is attached to - you cannot
attach a VM in one vDC to a disk in another.

This has been made a list-of-hashes specifically so that it can potentially be extended
later -- the API permits assigning independent disks to controller types and bus numbers.
Using hashes allows for this update without changing the API.