Update VM schema with independent disks; test.
I wanted to be sure that the schema addition for independent_disks worked, but
noticed that there wasn't any existing testing of the schema files.

I've added spec/vcloud/launcher/schema/launcher_vapps_spec.rb to test the particular
cases that I wanted to be sure worked as expected. It could be expanded to cover more
cases if required.