Fix the post-launch-cmd integration tests
Matt found that the tests fail because Vcloud::Core.logger is created
when first used. This means we need to set the expectations before we
first make a log message.

Additionally I found that I need to expect an additional debug message
about instatiating the vapp, otherwise the expectation to find a debug
message about running the script will fail when it sees the first log
message about instatiating the vapp.