Introduce a delay between Vapp and Disk delete
There appears to be a bug in vCD that does not fully register
Independent Disks as detached when a vApp is deleted. After a 15-20
second delay, the disk is marked detached as expected, but during
that period any attempt to remove it will fail with an 'disk is attached'
error.

NB: The #attached_vms call does not help - it returns empty immediately.