Explicitly set rbenv version 2.0.0
Explicitly set the rbenv version to 2.0.0, the oldest version we
support, in both `jenkins.sh` and `jenkins_tests.sh`.

Run the tests using the oldest version so as to increase the likelihood
that we'll catch changes that break backwards compatibility.

Use a `rbenv_version.sh` file to avoid duplicating the version number.
I've avoided using `.ruby-version` file, which is supported by rbenv, to
avoid the implication that the version mentioned in `.ruby-version` is
the only one supported by the project.